+++ servicekit/config.go
@@ -59,12 +59,12 @@ type ControlConfig struct {
 }
 
 type ServiceConfig struct {
-	Name               string `json:"name" yaml:"name"`
-	GRPCAddr           string `json:"grpc_addr" yaml:"grpc_addr"`
-	AdvertiseGRPCAddr  string `json:"advertise_grpc_addr" yaml:"advertise_grpc_addr"`
-	AdminAddr          string `json:"admin_addr" yaml:"admin_addr"`
-	AdvertiseAdminAddr string `json:"advertise_admin_addr" yaml:"advertise_admin_addr"`
-	EnablePprof        bool   `json:"enable_pprof" yaml:"enable_pprof"`
+	Name              string `json:"name" yaml:"name"`
+	GRPCAddr          string `json:"grpc_addr" yaml:"grpc_addr"`
+	AdvertiseGRPCAddr string `json:"advertise_grpc_addr" yaml:"advertise_grpc_addr"`
+	HTTPAddr          string `json:"http_addr" yaml:"http_addr"`
+	AdvertiseHTTPAddr string `json:"advertise_http_addr" yaml:"advertise_http_addr"`
+	EnablePprof       bool   `json:"enable_pprof" yaml:"enable_pprof"`
 }
 
 type RegistryConfig struct {